sqlite_db: changed last_updated to BigInt

+++ backupd/src/storage/sqlite_db/model.rs
@@ -1,5 +1,3 @@
-use std::time::Instant;
-
 use crate::storage::sqlite_db::schema::files;
 
 #[derive(Queryable, Insertable)]
@@ -7,5 +5,5 @@ use crate::storage::sqlite_db::schema::files;
 pub struct DbFile {
     pub real_filename: String,
     pub local_filename: String,
-    pub last_updated: i32,
+    pub last_updated: i64,
 }
