+++ backupd/src/storage/sqlite_db/schema.rs
@@ -2,6 +2,6 @@ table! {
     files (real_filename) {
         real_filename -> Text,
         local_filename -> Text,
-        last_updated -> Integer,
+        last_updated -> BigInt,
     }
 }
